requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
future==0.18.2
idna==2.8
line-bot-sdk==1.17.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.22.0
sqlparse==0.4.1
urllib3==1.25.11
